Move reporting to a service
